+++ python/test/dynamical_systems/test_ring.py
@@ -29,7 +29,7 @@ class TestRing(unittest.TestCase):
         center = sr.CartesianPose.Identity("CAttractor", "A")
 
         self.assertTrue(ds.get_parameter_value("center").is_empty())
-        ds.set_parameter(sr.Parameter("center", center, sr.StateType.PARAMETER_CARTESIANPOSE))
+        ds.set_parameter(sr.Parameter("center", center, sr.ParameterType.STATE, sr.StateType.CARTESIAN_POSE))
         self.assertFalse(ds.get_parameter_value("center").is_empty())
 
         self.assertFalse(ds.get_base_frame().is_empty())
@@ -56,7 +56,7 @@ class TestRing(unittest.TestCase):
         with self.assertRaises(RuntimeError):
             ds.evaluate(state4)
 
-        ds.set_parameter(sr.Parameter("center", center, sr.StateType.PARAMETER_CARTESIANPOSE))
+        ds.set_parameter(sr.Parameter("center", center, sr.ParameterType.STATE, sr.StateType.CARTESIAN_POSE))
         self.assertTrue(ds.is_compatible(state1))
         self.assertFalse(ds.is_compatible(state2))
         self.assertTrue(ds.is_compatible(state3))
@@ -64,10 +64,10 @@ class TestRing(unittest.TestCase):
 
     def test_points_on_radius(self):
         ds = CartesianRingDS()
-        ds.set_parameter(sr.Parameter("center", self.center, sr.StateType.PARAMETER_CARTESIANPOSE))
-        ds.set_parameter(sr.Parameter("radius", self.radius, sr.StateType.PARAMETER_DOUBLE))
-        ds.set_parameter(sr.Parameter("width", self.width, sr.StateType.PARAMETER_DOUBLE))
-        ds.set_parameter(sr.Parameter("speed", self.speed, sr.StateType.PARAMETER_DOUBLE))
+        ds.set_parameter(sr.Parameter("center", self.center, sr.ParameterType.STATE, sr.StateType.CARTESIAN_POSE))
+        ds.set_parameter(sr.Parameter("radius", self.radius, sr.ParameterType.DOUBLE))
+        ds.set_parameter(sr.Parameter("width", self.width, sr.ParameterType.DOUBLE))
+        ds.set_parameter(sr.Parameter("speed", self.speed, sr.ParameterType.DOUBLE))
 
         current_pose = copy.deepcopy(self.center)
         twist = sr.CartesianTwist(ds.evaluate(current_pose))
@@ -100,8 +100,8 @@ class TestRing(unittest.TestCase):
     def test_convergence_on_radius_random_center(self):
         center = sr.CartesianPose.Random("A")
         ds = CartesianRingDS()
-        ds.set_parameter(sr.Parameter("center", center, sr.StateType.PARAMETER_CARTESIANPOSE))
-        ds.set_parameter(sr.Parameter("radius", self.radius, sr.StateType.PARAMETER_DOUBLE))
+        ds.set_parameter(sr.Parameter("center", center, sr.ParameterType.STATE, sr.StateType.CARTESIAN_POSE))
+        ds.set_parameter(sr.Parameter("radius", self.radius, sr.ParameterType.DOUBLE))
 
         current_pose = sr.CartesianPose("B", self.radius * np.random.rand(3, 1))
         for i in range(self.nb_steps):