+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Release Versions:
 ## Upcoming changes (in development)
 
 - Fix if else conditions in setup.py to correctly install modules (#285)
-- Refactor state type (#277)
+- Refactor state type (#277, #278, #280)
 
 ## 5.2.0